Backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
pydantic>=2.8.0
yfinance>=0.2.50
pymongo>=4.6.0
python-dotenv>=1.0.1
